+++ packages/js-drive/test/unit/dpp/DriveStateRepository.spec.js
@@ -180,7 +180,7 @@ describe('DriveStateRepository', () => {
   describe('#storeDocument', () => {
     it('should store document in repository', async function it() {
       const storeMock = this.sinon.stub();
-      createDocumentRepositoryMock.returns({
+      createDocumentRepositoryMock.resolves({
         store: storeMock,
       });
 
@@ -202,7 +202,7 @@ describe('DriveStateRepository', () => {
       const type = 'documentType';
 
       const deleteMock = this.sinon.stub();
-      createDocumentRepositoryMock.returns({
+      createDocumentRepositoryMock.resolves({
         delete: deleteMock,
       });